Move PyVmomiDeviceHelper to module_utils.vmware_guest
